backend/requirements_advanced.txt
# Advanced AI Image Studio Backend Requirements
# ============================================

# Core FastAPI stack
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# Async file operations
aiofiles==23.2.1
aioredis==2.0.1

# AI/ML libraries
torch>=2.1.0
torchvision>=0.16.0
realesrgan==0.3.0
basicsr>=1.4.2
gfpgan>=1.3.8
opencv-python==4.8.1.78
Pillow>=10.0.0
numpy>=1.24.0

# Image processing
imageio>=2.33.1
scikit-image>=0.22.0

# Background task processing
celery[redis]==5.3.4
redis==5.0.1

# Data validation and serialization
pydantic>=2.0.0
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
isort==5.13.2
mypy==1.8.0

# Optional GPU acceleration
# Uncomment if using NVIDIA GPUs:
# nvidia-ml-py==12.535.133

# Optional face enhancement
# rembg>=2.0.50  # Background removal
# insightface>=0.7.3  # Face analysis